code/apps/kn-test/requirements.txt
# asyncio_mqtt==0.12.1
cloudevents
# Flask
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
httpx
logfmter
# pydantic==1.9.1
aiomqtt
python-ulid
Flask==2.0.3